Create Extracting Column from Table
In Ayehu NG, There exists a powerful and convenient activity for getting Row from the table(GetRows). This activity fetches a particular row from the table. But we don't have the same activity for Column many times we need to fetch the only column which can be achieved by using two activity RotateTable and GetRows. This tutorial details the proper usage of this activity for getting the column.
support article: https://support.ayehu.com/hc/en-us/articles/360049907034-Extracting-Column-from-Table